% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game_Pareto.R
\name{game_pareto}
\alias{game_pareto}
\title{Pareto Equilibrium}
\usage{
game_pareto(num_of_firms, K = "none")
}
\arguments{
\item{num_of_firms}{Number of firms in the game.}

\item{K}{Amount available of the Commons. Default is number of firms + 1.}
}
\value{
A written statement explaining the strategies that lead to pareto optimality.
}
\description{
Pareto Equilibrium
}
\examples{
game_pareto(2)
}
