% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game_Payoffs.R
\name{game_payoffs}
\alias{game_payoffs}
\title{Game Payoffs}
\usage{
game_payoffs(
  num_of_firms,
  p = 0.5,
  K = "none",
  S0 = 0,
  S1 = 1,
  S2 = 2,
  best_resp = "Yes"
)
}
\arguments{
\item{num_of_firms}{Number of firms in the game.}

\item{p}{penalty or entrance fee paid by the firms when entering the game, if they choose a strategy that is not S0. Default is 0.5.}

\item{K}{amount available of the Commons. Default is number of firms + 1.}

\item{S0}{Strategy of extracting 0 units or not submitting a form request to the government, in our game's context.}

\item{S1}{Strategy of extracting x unit from the Commons. Default is 1.}

\item{S2}{Strategy of extracting y units from the Commons. Default is 1.}

\item{best_resp}{if "Yes", it will return only a dataframe of the payoffs of Player i, for each strategy. Default at "Yes". If "No", it will print a phrase summarising the best responses and payoffs of Player i.}
}
\value{
payoffs of Player i and Player j
}
\description{
Game Payoffs
}
\examples{
game_payoffs(2)
}
