% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game_Simulations.R
\name{game_simulations}
\alias{game_simulations}
\title{Game Simulations}
\usage{
game_simulations(
  num_simulations,
  best_resp = "Yes",
  pareto = "Yes",
  NE = "Yes"
)
}
\arguments{
\item{num_simulations}{Maximum amount of players you want to visualize in the simulations -- the simulations start at 2. For example, if 6, then it will return the structure of the game for 2 players, later it will print the structure of the game for 3 players, later for 4 players and so on until it finishes by printing the structure of the game for 6 players.}

\item{best_resp}{If "Yes", returns payoffs for Player i, for each of their strategies. If "No", it will print a phrase summarising the best responses and payoffs of Player i.}

\item{pareto}{If "Yes", returns the pareto equilibrium strategies. Default is "Yes".}

\item{NE}{If "Yes", returns Nash Equilibrium strategies. Default is "Yes"}
}
\value{
Game structure for different number of players.
}
\description{
Game Simulations
}
\examples{
game_simulations(6)
}
